# python: build with --enable-optimizations and --with-lto