- [Y] Have you followed the [guidelines for contributing](https://github.com/Homebrew/homebrew-core/blob/HEAD/CONTRIBUTING.md)?
- [Y] Have you checked that there aren't other open [pull requests](https://github.com/Homebrew/homebrew-core/pulls) for the same formula update/change?
- [Y] Have you built your formula locally with `brew install --build-from-source <formula>`, where `<formula>` is the name of the formula you're submitting?
- [Y] Is your test running fine `brew test <formula>`, where `<formula>` is the name of the formula you're submitting?
- [Y] Does your build pass `brew audit --strict <formula>` (after doing `brew install <formula>`)?

-----

Recent python offers flags to enable PGO and LTO which hopefully results in a faster Python interpreter. This PR adds the required flags to the homebrew `configure` invocation. It also papers over a macOS specific issue where the `configure` script believes it needs `llvm-ar` when building `--with-lto`. I don't think that is actually true on macOS, as Xcode doesn't provide `llvm-ar`. I wasn't sure if there was an expansion like `#{MacOS.sdk_path}` that would be a better prefix than `/usr` when applying this override.