@dtrodrigues 2021-02-01
Some previous discussion regarding those flags:
https://github.com/Homebrew/homebrew-core/pull/45337#issuecomment-542072864 
https://mail.python.org/archives/list/python-dev@python.org/thread/EILILECNTLTW4VCBCPW37R4QRU7ZBDEU/#VUA2AG6J2QG2GK52DZ4VKZMRUTKJ4DWO