@acmorrow 2021-02-01
As suggested by a colleague who knows far more about python than I do, I ran `pyperformance` to compare the baseline `python@3.9` formula interpreter performance (compiled from source, not pulled from bottle) to that of one built from this PR. Results:

```
/usr/local/bin/python3.9 -m pyperformance compare ./homebrew-py39-baseline.json ./homebrew-py39-pr-70177.json 
homebrew-py39-baseline.json
===========================

Performance version: 1.0.1
Report on macOS-11.1-x86_64-i386-64bit
Number of logical CPUs: 8
Start date: 2021-02-01 16:42:05.251702
End date: 2021-02-01 17:12:14.829755

homebrew-py39-pr-70177.json
===========================

Performance version: 1.0.1
Report on macOS-11.1-x86_64-i386-64bit
Number of logical CPUs: 8
Start date: 2021-02-01 17:26:15.358418
End date: 2021-02-01 17:51:12.170954

### 2to3 ###
Mean +- std dev: 637 ms +- 35 ms -> 522 ms +- 18 ms: 1.22x faster
Significant (t=22.63)

### chameleon ###
Mean +- std dev: 19.2 ms +- 0.7 ms -> 14.5 ms +- 0.8 ms: 1.32x faster
Significant (t=31.88)

### chaos ###
Mean +- std dev: 233 ms +- 11 ms -> 160 ms +- 6 ms: 1.45x faster
Significant (t=43.79)

### crypto_pyaes ###
Mean +- std dev: 222 ms +- 17 ms -> 164 ms +- 7 ms: 1.36x faster
Significant (t=24.06)

### deltablue ###
Mean +- std dev: 15.9 ms +- 1.3 ms -> 10.8 ms +- 0.4 ms: 1.48x faster
Significant (t=29.20)

### django_template ###
Mean +- std dev: 107 ms +- 6 ms -> 76 ms +- 6 ms: 1.40x faster
Significant (t=26.72)

### dulwich_log ###
Mean +- std dev: 130 ms +- 11 ms -> 103 ms +- 6 ms: 1.26x faster
Significant (t=17.61)

### fannkuch ###
Mean +- std dev: 855 ms +- 19 ms -> 676 ms +- 11 ms: 1.26x faster
Significant (t=64.29)

### float ###
Mean +- std dev: 219 ms +- 6 ms -> 165 ms +- 6 ms: 1.32x faster
Significant (t=47.87)

### genshi_text ###
Mean +- std dev: 55.0 ms +- 2.2 ms -> 41.4 ms +- 1.5 ms: 1.33x faster
Significant (t=39.74)

### genshi_xml ###
Mean +- std dev: 119 ms +- 5 ms -> 92 ms +- 4 ms: 1.30x faster
Significant (t=34.09)

### go ###
Mean +- std dev: 521 ms +- 21 ms -> 362 ms +- 6 ms: 1.44x faster
Significant (t=56.76)

### hexiom ###
Mean +- std dev: 19.7 ms +- 0.8 ms -> 14.8 ms +- 0.9 ms: 1.33x faster
Significant (t=32.76)

### json_dumps ###
Mean +- std dev: 25.7 ms +- 1.0 ms -> 21.2 ms +- 1.2 ms: 1.21x faster
Significant (t=22.54)

### json_loads ###
Mean +- std dev: 51.9 us +- 2.6 us -> 39.4 us +- 2.3 us: 1.32x faster
Significant (t=27.56)

### logging_format ###
Mean +- std dev: 20.0 us +- 2.9 us -> 14.7 us +- 1.0 us: 1.36x faster
Significant (t=13.41)

### logging_silent ###
Mean +- std dev: 407 ns +- 32 ns -> 290 ns +- 16 ns: 1.40x faster
Significant (t=25.17)

### logging_simple ###
Mean +- std dev: 18.8 us +- 2.2 us -> 12.9 us +- 0.7 us: 1.46x faster
Significant (t=20.23)

### mako ###
Mean +- std dev: 33.8 ms +- 4.1 ms -> 24.9 ms +- 0.5 ms: 1.36x faster
Significant (t=16.64)

### meteor_contest ###
Mean +- std dev: 182 ms +- 12 ms -> 147 ms +- 4 ms: 1.24x faster
Significant (t=22.46)

### nbody ###
Mean +- std dev: 283 ms +- 18 ms -> 185 ms +- 5 ms: 1.53x faster
Significant (t=40.29)

### nqueens ###
Mean +- std dev: 194 ms +- 18 ms -> 160 ms +- 10 ms: 1.22x faster
Significant (t=13.07)

### pathlib ###
Mean +- std dev: 79.5 ms +- 5.5 ms -> 69.9 ms +- 4.6 ms: 1.14x faster
Significant (t=10.37)

### pickle ###
Mean +- std dev: 22.0 us +- 0.8 us -> 17.4 us +- 0.5 us: 1.27x faster
Significant (t=38.38)

### pickle_dict ###
Mean +- std dev: 48.0 us +- 1.6 us -> 36.6 us +- 1.2 us: 1.31x faster
Significant (t=43.32)

### pickle_list ###
Mean +- std dev: 7.54 us +- 0.40 us -> 5.54 us +- 0.15 us: 1.36x faster
Significant (t=36.64)

### pickle_pure_python ###
Mean +- std dev: 1.01 ms +- 0.05 ms -> 0.68 ms +- 0.02 ms: 1.49x faster
Significant (t=47.08)

### pidigits ###
Mean +- std dev: 264 ms +- 10 ms -> 268 ms +- 6 ms: 1.02x slower
Not significant

### pyflate ###
Mean +- std dev: 1.42 sec +- 0.04 sec -> 1.08 sec +- 0.05 sec: 1.31x faster
Significant (t=42.80)

### python_startup ###
Mean +- std dev: 29.7 ms +- 1.2 ms -> 29.2 ms +- 2.2 ms: 1.02x faster
Not significant

### python_startup_no_site ###
Mean +- std dev: 24.0 ms +- 1.2 ms -> 23.7 ms +- 1.8 ms: 1.01x faster
Not significant

### raytrace ###
Mean +- std dev: 1.11 sec +- 0.05 sec -> 0.72 sec +- 0.04 sec: 1.55x faster
Significant (t=45.60)

### regex_compile ###
Mean +- std dev: 349 ms +- 22 ms -> 253 ms +- 10 ms: 1.38x faster
Significant (t=30.42)

### regex_dna ###
Mean +- std dev: 275 ms +- 10 ms -> 247 ms +- 9 ms: 1.11x faster
Significant (t=15.74)

### regex_effbot ###
Mean +- std dev: 4.68 ms +- 0.18 ms -> 4.58 ms +- 0.33 ms: 1.02x faster
Significant (t=2.05)

### regex_v8 ###
Mean +- std dev: 37.6 ms +- 2.0 ms -> 33.1 ms +- 2.0 ms: 1.14x faster
Significant (t=12.22)

### richards ###
Mean +- std dev: 155 ms +- 7 ms -> 107 ms +- 7 ms: 1.44x faster
Significant (t=35.39)

### scimark_fft ###
Mean +- std dev: 752 ms +- 20 ms -> 530 ms +- 19 ms: 1.42x faster
Significant (t=62.60)

### scimark_lu ###
Mean +- std dev: 318 ms +- 9 ms -> 229 ms +- 6 ms: 1.39x faster
Significant (t=65.07)

### scimark_monte_carlo ###
Mean +- std dev: 218 ms +- 9 ms -> 153 ms +- 8 ms: 1.42x faster
Significant (t=41.11)

### scimark_sor ###
Mean +- std dev: 426 ms +- 18 ms -> 297 ms +- 9 ms: 1.43x faster
Significant (t=48.93)

### scimark_sparse_mat_mult ###
Mean +- std dev: 10.0 ms +- 0.7 ms -> 7.0 ms +- 0.4 ms: 1.43x faster
Significant (t=28.82)

### spectral_norm ###
Mean +- std dev: 309 ms +- 10 ms -> 213 ms +- 10 ms: 1.45x faster
Significant (t=52.20)

### sqlalchemy_declarative ###
Mean +- std dev: 271 ms +- 14 ms -> 233 ms +- 22 ms: 1.17x faster
Significant (t=11.61)

### sqlalchemy_imperative ###
Mean +- std dev: 49.8 ms +- 2.5 ms -> 42.3 ms +- 3.6 ms: 1.18x faster
Significant (t=13.26)

### sqlite_synth ###
Mean +- std dev: 4.91 us +- 0.24 us -> 4.25 us +- 0.29 us: 1.16x faster
Significant (t=13.57)

### sympy_expand ###
Mean +- std dev: 926 ms +- 42 ms -> 704 ms +- 29 ms: 1.31x faster
Significant (t=33.60)

### sympy_integrate ###
Mean +- std dev: 43.1 ms +- 2.5 ms -> 36.4 ms +- 4.3 ms: 1.19x faster
Significant (t=10.48)

### sympy_str ###
Mean +- std dev: 593 ms +- 21 ms -> 474 ms +- 28 ms: 1.25x faster
Significant (t=26.42)

### sympy_sum ###
Mean +- std dev: 385 ms +- 20 ms -> 310 ms +- 19 ms: 1.24x faster
Significant (t=20.90)

### telco ###
Mean +- std dev: 12.1 ms +- 0.6 ms -> 9.1 ms +- 0.7 ms: 1.33x faster
Significant (t=25.63)

### unpack_sequence ###
Mean +- std dev: 101 ns +- 4 ns -> 90 ns +- 5 ns: 1.12x faster
Significant (t=13.93)

### unpickle ###
Mean +- std dev: 27.9 us +- 1.3 us -> 21.5 us +- 1.2 us: 1.30x faster
Significant (t=27.73)

### unpickle_list ###
Mean +- std dev: 7.65 us +- 0.29 us -> 5.68 us +- 0.43 us: 1.35x faster
Significant (t=29.43)

### unpickle_pure_python ###
Mean +- std dev: 709 us +- 39 us -> 491 us +- 16 us: 1.44x faster
Significant (t=40.12)

### xml_etree_generate ###
Mean +- std dev: 163 ms +- 4 ms -> 132 ms +- 9 ms: 1.24x faster
Significant (t=25.11)

### xml_etree_iterparse ###
Mean +- std dev: 191 ms +- 7 ms -> 159 ms +- 6 ms: 1.20x faster
Significant (t=26.51)

### xml_etree_parse ###
Mean +- std dev: 261 ms +- 9 ms -> 238 ms +- 13 ms: 1.10x faster
Significant (t=11.37)

### xml_etree_process ###
Mean +- std dev: 130 ms +- 3 ms -> 105 ms +- 8 ms: 1.23x faster
Significant (t=22.03)

Skipped 1 benchmarks only in homebrew-py39-pr-70177.json: tornado_http
```